+++ src/test/static/namespaces/namespace-dynamic-unnamed.js
@@ -1,0 +1,14 @@
+/**
+ * @namespace
+ */
+DynamicNamespace['ArrayNotation'] = {
+  foo: 'bar'
+};
+
+var baz = 'abc';
+/**
+ * @namespace
+ */
+DynamicNamespace[baz] = {
+  foo: 'bar'
+};