+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 * Analyzer will not attempt to load or add warnings for imports which can not be resolved by its urlResolver.
 * Protocol-less URLs such as `//host/path` are properly handled instead of treated as absolute paths.
+* Add a new 'namespace' feature to support the `@namespace` jsdoc tag.
 
 ## [2.0.0-alpha.26] - 2017-02-22
 * Fix issue with file missing from package.json "files" array.